07.30 c

+++ hwCommander/src/main/java/com/hw/model/PartsGpuVO.java
@@ -22,7 +22,7 @@ public class PartsGpuVO {
 //	private String gpuasCdNm;
 	private String makerId;
 	private String makerNm;
-	private BigDecimal qc;
+	private int qc;
 	private int tdp;
 	private int bn;
 	private int il;
